Pair programming with @travs.  More factoring out of tests: 1) ConfigureTest 2) ConfigureableTest 3) DataObjectTestToo
