+++ tests/ConfigureableTest.py
@@ -1,0 +1,30 @@
+import sys
+sys.path.insert(0,".")
+import unittest
+import neuroml
+import neuroml.writers as writers
+import PyOpenWorm
+from PyOpenWorm import *
+import test_data as TD
+import networkx
+import rdflib
+import rdflib as R
+import pint as Q
+import os
+import subprocess as SP
+import subprocess
+import tempfile
+import doctest
+
+from glob import glob
+
+class ConfigureableTest(unittest.TestCase):
+    def test_init_empty(self):
+        """Ensure Configureable gets init'd with the defalut if nothing's given"""
+        i = Configureable()
+        self.assertEqual(Configureable.conf,i.conf)
+
+    def test_init_False(self):
+        """Ensure Configureable gets init'd with the defalut if False is given"""
+        i = Configureable(conf=False)
+        self.assertEqual(Configureable.conf, i.conf)
