+++ tests/test.py
@@ -62,81 +62,14 @@ from ExampleRunnerTest import ExampleRunnerTest
 # Need description of these tests
 from DataIntegrityTest import DataIntegrityTest
 
-class ConfigureTest(unittest.TestCase):
-    def test_fake_config(self):
-        """ Try to retrieve a config value that hasn't been set """
-        with self.assertRaises(KeyError):
-            c = Configure()
-            c['not_a_valid_config']
+# Need description of these tests
+from ConfigureTest import ConfigureTest
 
-    def test_literal(self):
-        """ Assign a literal rather than a ConfigValue"""
-        c = Configure()
-        c['seven'] = "coke"
-        self.assertEqual(c['seven'], "coke")
+# Need description of these tests
+from ConfigureableTest import ConfigureableTest
 
-    def test_ConfigValue(self):
-        """ Assign a ConfigValue"""
-        c = Configure()
-        class pipe(ConfigValue):
-            def get(self):
-                return "sign"
-        c['seven'] = pipe()
-        self.assertEqual("sign",c['seven'])
-
-    def test_getter_no_ConfigValue(self):
-        """ Assign a method with a "get". Should return a the object rather than calling its get method """
-        c = Configure()
-        class pipe:
-            def get(self):
-                return "sign"
-        c['seven'] = pipe()
-        self.assertIsInstance(c['seven'], pipe)
-
-    def test_late_get(self):
-        """ "get" shouldn't be called until the value is *dereferenced* """
-        c = Configure()
-        a = {'t' : False}
-        class pipe(ConfigValue):
-            def get(self):
-                a['t'] = True
-                return "sign"
-        c['seven'] = pipe()
-        self.assertFalse(a['t'])
-        self.assertEqual(c['seven'], "sign")
-        self.assertTrue(a['t'])
-
-    def test_read_from_file(self):
-        """ Read configuration from a JSON file """
-        try:
-            d = Data.open("tests/test.conf")
-            self.assertEqual("test_value", d["test_variable"])
-        except:
-            self.fail("test.conf should exist and be valid JSON")
-
-    def test_read_from_file_fail(self):
-        """ Fail on attempt to read configuration from a non-JSON file """
-        with self.assertRaises(ValueError):
-            Data.open("tests/bad_test.conf")
-
-
-class ConfigureableTest(unittest.TestCase):
-    def test_init_empty(self):
-        """Ensure Configureable gets init'd with the defalut if nothing's given"""
-        i = Configureable()
-        self.assertEqual(Configureable.conf,i.conf)
-
-    def test_init_False(self):
-        """Ensure Configureable gets init'd with the defalut if False is given"""
-        i = Configureable(conf=False)
-        self.assertEqual(Configureable.conf, i.conf)
-
-class DataObjectTestToo(unittest.TestCase):
-    def test_helpful_message_on_non_connection(self):
-        """ The message should say something about connecting """
-        Configureable.conf = False # Ensure that we are disconnected
-        with self.assertRaisesRegexp(Exception, ".*[cC]onnect.*"):
-            do = DataObject()
+# Need description of these tests
+from DataObjectTestToo import DataObjectTestToo
 
 class _DataTest(unittest.TestCase):
     def delete_dir(self):
